bottypes/ for bot-specific stuff, BClient -> BotClient
