+++ main.py
@@ -16,19 +16,18 @@ from pyrogram.types import CallbackQuery, Message
 from pyropatch import pyropatch  # do not delete!!
 from telegraph.aio import Telegraph
 
+from bottypes import BotClient, ExtendedIKB, ExtendedIKM, UserSession
 import config
 from db import db_session
 from functions import datacenter_handlers, info_formatters
 from functions.decorators import ignore_message_not_modified
 from functions.locale import get_available_languages
 import keyboards
-from keyboards import ExtendedIKB, ExtendedIKM
 # noinspection PyPep8Naming
 from l10n import Locale, LocaleKeys as LK
-from utypes import (BClient, ExchangeRate, GameServersData,
-                    GameVersionData, GunInfo, LeaderboardStats,
-                    ProfileInfo, State, States, UserGameStats,
-                    UserSession, drop_cap_reset_timer)
+from utypes import (ExchangeRate, GameServersData, GameVersionData,
+                    GunInfo, LeaderboardStats, ProfileInfo, State,
+                    States, UserGameStats, drop_cap_reset_timer)
 from utypes.profiles import ErrorCode, ParseUserStatsError  # to clearly indicate relation
 
 
@@ -40,15 +39,15 @@ logging.basicConfig(level=logging.INFO,
                     format="%(asctime)s | %(threadName)s: %(message)s",
                     datefmt="%H:%M:%S — %d/%m/%Y")
 
-bot = BClient(config.BOT_NAME,
-              api_id=config.API_ID,
-              api_hash=config.API_HASH,
-              bot_token=config.BOT_TOKEN,
-              plugins={'root': 'plugins'},
-              test_mode=config.TEST_MODE or True,
-              workdir=config.SESS_FOLDER,
-              log_channel=config.LOGCHANNEL,
-              back_callback=LK.bot_back,)
+bot = BotClient(config.BOT_NAME,
+                api_id=config.API_ID,
+                api_hash=config.API_HASH,
+                bot_token=config.BOT_TOKEN,
+                plugins={'root': 'plugins'},
+                test_mode=config.TEST_MODE,
+                workdir=config.SESS_FOLDER,
+                log_channel=config.LOGCHANNEL,
+                back_callback=LK.bot_back,)
 
 telegraph = Telegraph(access_token=config.TELEGRAPH_ACCESS_TOKEN)
 
@@ -56,7 +55,7 @@ telegraph = Telegraph(access_token=config.TELEGRAPH_ACCESS_TOKEN)
 # cat: Main
 
 @bot.on_callback_exception()
-async def handle_exceptions_in_callback(client: BClient, session: UserSession, callback_query: CallbackQuery,
+async def handle_exceptions_in_callback(client: BotClient, session: UserSession, callback_query: CallbackQuery,
                                         exc: Exception):
     logging.exception('Caught exception!', exc_info=exc)
     await client.log(f'❗️ {traceback.format_tb(exc.__traceback__)}',
@@ -66,7 +65,7 @@ async def handle_exceptions_in_callback(client: BClient, session: UserSession, c
 
 
 @bot.on_message(~filters.me)
-async def sync_user_data(client: BClient, message: Message):
+async def sync_user_data(client: BotClient, message: Message):
     if message.chat.type != ChatType.PRIVATE:
         message.continue_propagation()
 
@@ -82,7 +81,7 @@ async def sync_user_data(client: BClient, message: Message):
 
 
 @bot.on_message(filters.command(ALL_COMMANDS))
-async def any_command(client: BClient, message: Message):
+async def any_command(client: BotClient, message: Message):
     await client.send_chat_action(message.chat.id, ChatAction.TYPING)
 
     user = message.from_user
@@ -96,7 +95,7 @@ async def any_command(client: BClient, message: Message):
 
 
 @bot.on_callback_query()
-async def sync_user_data_callback(client: BClient, callback_query: CallbackQuery):
+async def sync_user_data_callback(client: BotClient, callback_query: CallbackQuery):
     if callback_query.message.chat.type != ChatType.PRIVATE:
         return
 
@@ -140,7 +139,7 @@ async def server_stats(_, session: UserSession, callback_query: CallbackQuery):
 
 
 @bot.funcmenu(LK.game_status_button_title, came_from=server_stats, ignore_message_not_modified=True)
-async def send_server_status(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_server_status(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     """Send the status of Counter-Strike servers"""
 
     data = GameServersData.cached_server_status()
@@ -154,7 +153,7 @@ async def send_server_status(client: BClient, session: UserSession, callback_que
 
 
 @bot.funcmenu(LK.stats_matchmaking_button_title, came_from=server_stats, ignore_message_not_modified=True)
-async def send_matchmaking_stats(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_matchmaking_stats(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     """Send Counter-Strike matchamaking statistics"""
 
     data = GameServersData.cached_matchmaking_stats()
@@ -177,12 +176,12 @@ async def datacenters(_, session: UserSession, callback_query: CallbackQuery):
 
 
 @bot.funcmenu(LK.regions_africa, came_from=datacenters)
-async def send_dc_africa(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_africa(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.africa, keyboards.dc_markup)
 
 
 @bot.funcmenu(LK.regions_australia, came_from=datacenters)
-async def send_dc_australia(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_australia(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.australia, keyboards.dc_markup)
 
 
@@ -193,37 +192,37 @@ async def dc_europe(_, session: UserSession, callback_query: CallbackQuery):
 
 
 @bot.funcmenu(LK.dc_austria, came_from=dc_europe)
-async def send_dc_austria(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_austria(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.austria, keyboards.dc_eu_markup)
 
 
 @bot.funcmenu(LK.dc_finland, came_from=dc_europe)
-async def send_dc_finland(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_finland(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.finland, keyboards.dc_eu_markup)
 
 
 @bot.funcmenu(LK.dc_germany, came_from=dc_europe)
-async def send_dc_germany(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_germany(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.germany, keyboards.dc_eu_markup)
 
 
 @bot.funcmenu(LK.dc_netherlands, came_from=dc_europe)
-async def send_dc_netherlands(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_netherlands(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.netherlands, keyboards.dc_eu_markup)
 
 
 @bot.funcmenu(LK.dc_poland, came_from=dc_europe)
-async def send_dc_poland(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_poland(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.poland, keyboards.dc_eu_markup)
 
 
 @bot.funcmenu(LK.dc_spain, came_from=dc_europe)
-async def send_dc_spain(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_spain(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.spain, keyboards.dc_eu_markup)
 
 
 @bot.funcmenu(LK.dc_sweden, came_from=dc_europe)
-async def send_dc_sweden(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_sweden(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.sweden, keyboards.dc_eu_markup)
 
 
@@ -234,12 +233,12 @@ async def dc_us(_, session: UserSession, callback_query: CallbackQuery):
 
 
 @bot.funcmenu(LK.dc_us_north, came_from=dc_us)
-async def send_dc_us_north(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_us_north(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.us_north, keyboards.dc_us_markup)
 
 
 @bot.funcmenu(LK.dc_us_south, came_from=dc_us)
-async def send_dc_us_south(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_us_south(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.us_south, keyboards.dc_us_markup)
 
 
@@ -250,23 +249,23 @@ async def dc_southamerica(_, session: UserSession, callback_query: CallbackQuery
 
 
 @bot.funcmenu(LK.dc_argentina, came_from=dc_southamerica)
-async def send_dc_argentina(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_argentina(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query,
                         datacenter_handlers.argentina, keyboards.dc_southamerica_markup)
 
 
 @bot.funcmenu(LK.dc_brazil, came_from=dc_southamerica)
-async def send_dc_brazil(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_brazil(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.brazil, keyboards.dc_southamerica_markup)
 
 
 @bot.funcmenu(LK.dc_chile, came_from=dc_southamerica)
-async def send_dc_chile(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_chile(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.chile, keyboards.dc_southamerica_markup)
 
 
 @bot.funcmenu(LK.dc_peru, came_from=dc_southamerica)
-async def send_dc_peru(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_peru(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.peru, keyboards.dc_southamerica_markup)
 
 
@@ -277,42 +276,42 @@ async def dc_asia(_, session: UserSession, callback_query: CallbackQuery):
 
 
 @bot.funcmenu(LK.dc_india, came_from=dc_asia)
-async def send_dc_india(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_india(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.india, keyboards.dc_asia_markup)
 
 
 @bot.funcmenu(LK.dc_japan, came_from=dc_asia)
-async def send_dc_japan(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_japan(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.japan, keyboards.dc_asia_markup)
 
 
 @bot.funcmenu(LK.regions_china, came_from=dc_asia)
-async def send_dc_china(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_china(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.china, keyboards.dc_asia_markup)
 
 
 @bot.funcmenu(LK.dc_emirates, came_from=dc_asia)
-async def send_dc_emirates(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_emirates(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.emirates, keyboards.dc_asia_markup)
 
 
 @bot.funcmenu(LK.dc_singapore, came_from=dc_asia)
-async def send_dc_singapore(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_singapore(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.singapore, keyboards.dc_asia_markup)
 
 
 @bot.funcmenu(LK.dc_hongkong, came_from=dc_asia)
-async def send_dc_hongkong(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_hongkong(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.hongkong, keyboards.dc_asia_markup)
 
 
 @bot.funcmenu(LK.dc_southkorea, came_from=dc_asia)
-async def send_dc_south_korea(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_dc_south_korea(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_dc_state(client, session, callback_query, datacenter_handlers.south_korea, keyboards.dc_asia_markup)
 
 
 @ignore_message_not_modified
-async def send_dc_state(client: BClient, session: UserSession, callback_query: CallbackQuery,
+async def send_dc_state(client: BotClient, session: UserSession, callback_query: CallbackQuery,
                         dc_state_func: Callable[[Locale], str | State], reply_markup: ExtendedIKM):
     try:
         state = dc_state_func(session.locale)
@@ -328,7 +327,7 @@ async def send_dc_state(client: BClient, session: UserSession, callback_query: C
 
 
 @bot.navmenu(LK.bot_profile_info, came_from=main_menu, ignore_message_not_modified=True)
-async def profile_info(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def profile_info(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     with open(config.CACHE_FILE_PATH, encoding='utf-8') as f:
         cache_file = json.load(f)
 
@@ -340,7 +339,7 @@ async def profile_info(client: BClient, session: UserSession, callback_query: Ca
 
 
 @bot.funcmenu(LK.user_profileinfo_title, came_from=profile_info)
-async def user_profile_info(client: BClient, session: UserSession,
+async def user_profile_info(client: BotClient, session: UserSession,
                             callback_query: CallbackQuery, last_error: str = None):
     text = session.locale.steam_url_example if last_error is None else last_error
     text += '\n\n' + session.locale.bot_use_cancel
@@ -407,7 +406,7 @@ async def user_profile_info(client: BClient, session: UserSession,
 
 
 @bot.funcmenu(LK.user_gamestats_button_title, came_from=profile_info, ignore_message_not_modified=True)
-async def user_game_stats(client: BClient, session: UserSession, callback_query: CallbackQuery, last_error: str = None):
+async def user_game_stats(client: BotClient, session: UserSession, callback_query: CallbackQuery, last_error: str = None):
     text = session.locale.steam_url_example if last_error is None else last_error
     text += '\n\n' + session.locale.bot_use_cancel
 
@@ -491,7 +490,7 @@ async def generate_crosshair(_, session: UserSession, callback_query: CallbackQu
 
 
 @bot.funcmenu(LK.crosshair_decode, came_from=crosshair, ignore_message_not_modified=True)
-async def decode_crosshair(client: BClient, session: UserSession,
+async def decode_crosshair(client: BotClient, session: UserSession,
                            callback_query: CallbackQuery, last_error: str = None):
     text = session.locale.crosshair_decode_example if last_error is None else last_error
     text += '\n\n' + session.locale.bot_use_cancel
@@ -545,7 +544,7 @@ async def send_dropcap_timer(_, session: UserSession, callback_query: CallbackQu
 
 
 @bot.funcmenu(LK.game_version_button_title, came_from=extra_features, ignore_message_not_modified=True)
-async def send_game_version(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def send_game_version(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     """Send a current version of CS:GO/CS 2"""
 
     data = GameVersionData.cached_data()
@@ -568,42 +567,42 @@ async def game_leaderboard(_, session: UserSession, callback_query: CallbackQuer
 
 
 @bot.funcmenu(LK.game_leaderboard_world, came_from=game_leaderboard)
-async def game_leaderboard_world(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_world(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.game_leaderboard_world)
 
 
 @bot.funcmenu(LK.regions_africa, came_from=game_leaderboard)
-async def game_leaderboard_africa(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_africa(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.regions_africa)
 
 
 @bot.funcmenu(LK.regions_asia, came_from=game_leaderboard)
-async def game_leaderboard_asia(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_asia(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.regions_asia)
 
 
 @bot.funcmenu(LK.regions_australia, came_from=game_leaderboard)
-async def game_leaderboard_australia(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_australia(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.regions_australia)
 
 
 @bot.funcmenu(LK.regions_china, came_from=game_leaderboard)
-async def game_leaderboard_china(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_china(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.regions_china)
 
 
 @bot.funcmenu(LK.regions_europe, came_from=game_leaderboard)
-async def game_leaderboard_europe(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_europe(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.regions_europe)
 
 
 @bot.funcmenu(LK.regions_northamerica, came_from=game_leaderboard)
-async def game_leaderboard_northamerica(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_northamerica(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.regions_northamerica)
 
 
 @bot.funcmenu(LK.regions_southamerica, came_from=game_leaderboard)
-async def game_leaderboard_southamerica(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def game_leaderboard_southamerica(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await send_game_leaderboard(client, session, callback_query, LK.regions_southamerica)
 
 
@@ -638,7 +637,7 @@ async def guns(_, session: UserSession, callback_query: CallbackQuery):
 
 
 @bot.navmenu(LK.gun_pistols, came_from=guns)
-async def pistols(client: BClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
+async def pistols(client: BotClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
     if loop:
         chosen_gun = await client.listen_callback(callback_query.message.chat.id,
                                                   callback_query.message.id)
@@ -651,7 +650,7 @@ async def pistols(client: BClient, session: UserSession, callback_query: Callbac
 
 
 @bot.callback_process(of=pistols)
-async def pistols_process(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def pistols_process(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await client.log_callback(session, callback_query)
 
     chosen_gun = callback_query.data
@@ -666,7 +665,7 @@ async def pistols_process(client: BClient, session: UserSession, callback_query:
 
 
 @bot.navmenu(LK.gun_heavy, came_from=guns)
-async def heavy(client: BClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
+async def heavy(client: BotClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
     if loop:
         chosen_gun = await client.listen_callback(callback_query.message.chat.id,
                                                   callback_query.message.id)
@@ -679,7 +678,7 @@ async def heavy(client: BClient, session: UserSession, callback_query: CallbackQ
 
 
 @bot.callback_process(of=heavy)
-async def heavy_process(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def heavy_process(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await client.log_callback(session,  callback_query)
 
     chosen_gun = callback_query.data
@@ -694,7 +693,7 @@ async def heavy_process(client: BClient, session: UserSession, callback_query: C
 
 
 @bot.navmenu(LK.gun_smgs, came_from=guns)
-async def smgs(client: BClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
+async def smgs(client: BotClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
     if loop:
         chosen_gun = await client.listen_callback(callback_query.message.chat.id,
                                                   callback_query.message.id)
@@ -707,7 +706,7 @@ async def smgs(client: BClient, session: UserSession, callback_query: CallbackQu
 
 
 @bot.callback_process(of=smgs)
-async def smgs_process(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def smgs_process(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await client.log_callback(session, callback_query)
 
     chosen_gun = callback_query.data
@@ -721,7 +720,7 @@ async def smgs_process(client: BClient, session: UserSession, callback_query: Ca
 
 
 @bot.navmenu(LK.gun_rifles, came_from=guns)
-async def rifles(client: BClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
+async def rifles(client: BotClient, session: UserSession, callback_query: CallbackQuery, loop: bool = False):
     if loop:
         chosen_gun = await client.listen_callback(callback_query.message.chat.id,
                                                   callback_query.message.id)
@@ -734,7 +733,7 @@ async def rifles(client: BClient, session: UserSession, callback_query: Callback
 
 
 @bot.callback_process(of=rifles)
-async def rifles_process(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def rifles_process(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await client.log_callback(session, callback_query)
 
     chosen_gun = callback_query.data
@@ -747,7 +746,7 @@ async def rifles_process(client: BClient, session: UserSession, callback_query: 
     return await unknown_request(client, session, callback_query, keyboards.rifles_markup)
 
 
-async def send_gun_info(client: BClient, session: UserSession, callback_query: CallbackQuery, _from: callable,
+async def send_gun_info(client: BotClient, session: UserSession, callback_query: CallbackQuery, _from: callable,
                         gun_info: GunInfo, reply_markup: ExtendedIKM):
     """Send archived data about guns"""
 
@@ -778,7 +777,7 @@ async def settings(_, session: UserSession, callback_query: CallbackQuery):
 
 
 @bot.navmenu(LK.settings_language_button_title, came_from=settings, ignore_message_not_modified=True)
-async def language(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def language(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     keyboards.language_settings_markup.select_button_by_key(session.locale.lang_code)
 
     chosen_lang = await client.ask_callback_silently(
@@ -791,7 +790,7 @@ async def language(client: BClient, session: UserSession, callback_query: Callba
 
 
 @bot.callback_process(of=language)
-async def language_process(client: BClient, session: UserSession, callback_query: CallbackQuery):
+async def language_process(client: BotClient, session: UserSession, callback_query: CallbackQuery):
     await client.log_callback(session, callback_query)
 
     chosen_lang = callback_query.data
@@ -806,7 +805,7 @@ async def language_process(client: BClient, session: UserSession, callback_query
 
 
 @bot.on_command('start')
-async def welcome(client: BClient, session: UserSession, message: Message):
+async def welcome(client: BotClient, session: UserSession, message: Message):
     """First bot's message"""
 
     if message.chat.type != ChatType.PRIVATE:
@@ -821,7 +820,7 @@ async def welcome(client: BClient, session: UserSession, message: Message):
 
 
 @bot.on_command('feedback')
-async def leave_feedback(client: BClient, session: UserSession, message: Message):
+async def leave_feedback(client: BotClient, session: UserSession, message: Message):
     """Send feedback"""
 
     if message.chat.type != ChatType.PRIVATE:
@@ -848,7 +847,7 @@ async def leave_feedback(client: BClient, session: UserSession, message: Message
 
 
 @bot.on_command('help')
-async def _help(client: BClient, session: UserSession, message: Message):
+async def _help(client: BotClient, session: UserSession, message: Message):
     """/help message"""
 
     if message.chat.type != ChatType.PRIVATE:
@@ -904,7 +903,7 @@ async def log_ping(_, __, callback_query: CallbackQuery):
     await callback_query.answer('Yes, I AM working!')
 
 
-async def regular_stats_report(client: BClient):
+async def regular_stats_report(client: BotClient):
     now = dt.datetime.now(dt.UTC)
 
     text = (f'📊 **Some stats for the past 8 hours:**\n'
